+++ src/mame/machine/docastle.c
@@ -36,14 +36,14 @@ if it was a small shared buffer. The order of operations is:
 */
 READ8_HANDLER( docastle_shared0_r )
 {
-	if (offset == 8) logerror("CPU #0 shared0r  clock = %d\n",activecpu_gettotalcycles());
+	if (offset == 8) logerror("CPU #0 shared0r  clock = %d\n",(UINT32)activecpu_gettotalcycles());
 	return buffer0[offset];
 }
 
 
 READ8_HANDLER( docastle_shared1_r )
 {
-	if (offset == 8) logerror("CPU #1 shared1r  clock = %d\n",activecpu_gettotalcycles());
+	if (offset == 8) logerror("CPU #1 shared1r  clock = %d\n",(UINT32)activecpu_gettotalcycles());
 	return buffer1[offset];
 }
 
@@ -51,7 +51,7 @@ READ8_HANDLER( docastle_shared1_r )
 WRITE8_HANDLER( docastle_shared0_w )
 {
 	if (offset == 8) logerror("CPU #1 shared0w %02x %02x %02x %02x %02x %02x %02x %02x %02x clock = %d\n",
-		buffer0[0],buffer0[1],buffer0[2],buffer0[3],buffer0[4],buffer0[5],buffer0[6],buffer0[7],data,activecpu_gettotalcycles());
+		buffer0[0],buffer0[1],buffer0[2],buffer0[3],buffer0[4],buffer0[5],buffer0[6],buffer0[7],data,(UINT32)activecpu_gettotalcycles());
 
 	buffer0[offset] = data;
 
@@ -68,7 +68,7 @@ WRITE8_HANDLER( docastle_shared1_w )
 	if (offset == 8)
 	{
 		logerror("CPU #0 shared1w %02x %02x %02x %02x %02x %02x %02x %02x %02x clock = %d\n",
-				buffer1[0],buffer1[1],buffer1[2],buffer1[3],buffer1[4],buffer1[5],buffer1[6],buffer1[7],data,activecpu_gettotalcycles());
+				buffer1[0],buffer1[1],buffer1[2],buffer1[3],buffer1[4],buffer1[5],buffer1[6],buffer1[7],data,(UINT32)activecpu_gettotalcycles());
 
 		/* freeze execution of the master CPU until the slave has used the shared memory */
 		cpu_spinuntil_trigger(500);
